pyproject.toml
[project]
name = "rss-stream-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "clickhouse-connect>=0.10.0",
    "feedparser>=6.0.12",
    "httpx[http2]>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]
